ISIS-2648: move mm mementos up to core
